Crate riddle_common

Riddle crate with miscelanious support functionality required by the rest of the riddle crates.

Modules

eventpub
An event pubsub utility.

Macros

define_handles
Implement CloneHandle trait and define handle types for a given object.

Structs

Color
An RGBA color, parameterized over color channel type.

Enums

CommonError
Errors common to many riddle crates

Traits

CloneHandle
Trait for objects that contain a weak reference to themselves, to allow functions which get references to those objects to clone handles to those objects.

ColorElement
A type that represents a single channel in a color value.

ColorElementConversion
Define the mapping between two ColorElement types.